# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "مستندات API سامانه لد ",
        "description": "توضیحات نحوه استفاده از APIها",
        "contact": {
            "email": "hreza779@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "سرور اصلی API"
        }
    ],
    "paths": {
...
        "/api/exam-sets": {
            "get": {
                "tags": [
                    "Exam Sets"
                ],
                "summary": "Get all exam sets",
                "operationId": "0bc0bf34d88c41db9a86f0751e5d4498",
                "parameters": [
                    {
                        "name": "manager_id",
                        "in": "query",
                        "description": "Filter by manager ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "in_progress",
                                "completed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of exam sets"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Exam Sets"
                ],
                "summary": "Create a new exam set",
                "operationId": "ff08d8ea4390ca06f842576ac7f69066",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "manager_id",
                                    "title",
                                    "exam_ids"
                                ],
                                "properties": {
                                    "manager_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "assessment_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Onboarding Exams"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "exam_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    },
                                    "assigned_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "exam_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exam set created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exam-sets/{examSet}": {
            "get": {
                "tags": [
                    "Exam Sets"
                ],
                "summary": "Get exam set details",
                "operationId": "00e7c802d79912cda6be473fbb2ee6f9",
                "parameters": [
                    {
                        "name": "examSet",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam set details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Exam Sets"
                ],
                "summary": "Update exam set",
                "operationId": "556cab365ad5aeb36e4af7e2391ada37",
                "parameters": [
                    {
                        "name": "examSet",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "exam_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exam set updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Exam Sets"
                ],
                "summary": "Delete exam set",
                "operationId": "76f0683673ed529f6a1b63f8b5e80185",
                "parameters": [
                    {
                        "name": "examSet",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Exam set deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exam-sets/{examSet}/exams": {
            "post": {
                "tags": [
                    "Exam Sets"
                ],
                "summary": "Add exams to exam set",
                "operationId": "70927492bfd93ddcf4a7d26d9cb4f2c4",
                "parameters": [
                    {
                        "name": "examSet",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "exam_ids"
                                ],
                                "properties": {
                                    "exam_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exams added successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}